+++ src/rust/src/declarative_asn1/types.rs
@@ -82,6 +82,8 @@ pub enum Type {
     CertificateSigningRequest(),
     /// `x509.CertificateRevocationList`
     CertificateRevocationList(),
+    /// `x509.Name`
+    Name(),
 }
 
 /// A type that we know how to encode/decode, along with any
@@ -579,6 +581,8 @@ pub fn non_root_python_to_rust<'p>(
         Type::CertificateSigningRequest().into_pyobject(py)
     } else if class.is(crate::x509::crl::CertificateRevocationList::type_object(py)) {
         Type::CertificateRevocationList().into_pyobject(py)
+    } else if class.is(&crate::types::NAME.get(py)?) {
+        Type::Name().into_pyobject(py)
     } else {
         Err(pyo3::exceptions::PyTypeError::new_err(format!(
             "cannot handle type: {class:?}"
@@ -723,12 +727,11 @@ pub(crate) fn is_tag_valid_for_type(
             }
         }
         Type::Null() => check_tag_with_encoding(asn1::Null::TAG, encoding, tag),
-        // Certificates, CSRs, and CRLs are all SEQUENCEs
+        // Certificates, CSRs, CRLs, and Names are all SEQUENCEs
         Type::Certificate()
         | Type::CertificateSigningRequest()
-        | Type::CertificateRevocationList() => {
-            check_tag_with_encoding(asn1::Sequence::TAG, encoding, tag)
-        }
+        | Type::CertificateRevocationList()
+        | Type::Name() => check_tag_with_encoding(asn1::Sequence::TAG, encoding, tag),
     }
 }